Second stage of fixes for custom experts
- Moved all existing layers on this API
- Fix some typos and beter namings
- Moved file with class_registrator to server.layers
- Added some tests, but they are marked to be skipped due to a mistake in backgrond_server
