+++ hivemind/server/expert_backend.py
@@ -79,6 +79,9 @@ class ExpertBackend(nn.Module):
         """
         args, kwargs = nested_pack(inputs, structure=self.forward_schema)
 
+        if args[0].shape[0] == 0:
+            raise RuntimeError("Batch should contain more than 0 samples")
+
         with torch.no_grad():
             outputs = self.expert(*args, **kwargs)
